Bug: 12524 Only remove events created by the user.

+++ kronolith/lib/Kronolith.php
@@ -2917,7 +2917,9 @@ class Kronolith
             }
             foreach ($uids as $uid) {
                 $event = $driver->getByUID($uid, array($calendar));
-                $driver->deleteEvent($event->id);
+                if ($event->creator == $user) {
+                    $driver->deleteEvent($event->id);
+                }
             }
         }
     }
